backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
pytest==8.3.4
httpx==0.28.1
Pillow==11.1.0
requests==2.32.3
